requirements.txt

nltk
flask 
pillow
cv2
flask-cors
docx2pdf
numpy
pandas
sklearn
matplotlib
wordcloud
urlextract
seaborn 
pypdf
pythoncom
pdf2docx
pdf2image
fitz  
zipfile